tensquare_1103

+++ tensquare_user/src/main/java/com/tensquare/user/controller/UserController.java
@@ -1,7 +1,8 @@
 package com.tensquare.user.controller;
-import java.util.List;
+import java.util.HashMap;
 import java.util.Map;
 
+import io.jsonwebtoken.Claims;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.redis.core.RedisTemplate;
@@ -18,6 +19,10 @@ import com.tensquare.user.service.UserService;
 import entity.PageResult;
 import entity.Result;
 import entity.StatusCode;
+import util.JwtUtil;
+
+import javax.servlet.http.HttpServletRequest;
+
 /**
  * 控制器层
  * @author Administrator
@@ -34,13 +39,34 @@ public class UserController {
 	@Autowired
 	private RedisTemplate redisTemplate;
 
+	@Autowired
+	private JwtUtil jwtUtil;
+
+	@Autowired
+	private HttpServletRequest request;
+
+	/**
+	 * 更新好友粉丝数和用户关注数
+	 * @param x
+	 * @param userid
+	 * @param friendid
+	 */
+	@RequestMapping(value = "/{userid}/{friendid}/{x}", method = RequestMethod.PUT)
+	public void updatefanscountandfollowscount(@PathVariable String userid, @PathVariable String friendid, @PathVariable int x) {
+		userService.updatefanscountandfollowcount(x, userid, friendid);
+	}
+
 	@RequestMapping(value = "/login", method = RequestMethod.POST)
 	public Result login(@RequestBody User user) {
 		user = userService.login(user.getMobile(), user.getPassword());
 		if (user == null) {
 			return new Result(false, StatusCode.LOGINERROR, "登陆失败");
 		}
-		return new Result(true, StatusCode.OK, "登陆成功");
+		String token = jwtUtil.createJWT(user.getId(), user.getMobile(), "user");
+		Map<String, Object> map = new HashMap<>();
+		map.put("token", token);
+		map.put("roles","user");
+		return new Result(true, StatusCode.OK, "登陆成功", map);
 	}
 
 	/**
@@ -142,8 +168,12 @@ public class UserController {
 	 */
 	@RequestMapping(value="/{id}",method= RequestMethod.DELETE)
 	public Result delete(@PathVariable String id ){
+		String token = (String) request.getAttribute("claims_admin");
+		if (token == null || "".equals(token)) {
+			return new Result(false, StatusCode.ACCESSERROR,"权限不足");
+		}
 		userService.deleteById(id);
-		return new Result(true,StatusCode.OK,"删除成功");
+		return new Result(true, StatusCode.OK,"删除成功");
 	}
 	
 }
